LPS-56513 Move to bundle
