+++ modules/portal/portal-cache/src/com/liferay/portal/cache/internal/MultiVMPoolImpl.java
@@ -12,11 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.cache;
+package com.liferay.portal.cache.internal;
 
+import com.liferay.portal.kernel.cache.MultiVMPool;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
-import com.liferay.portal.kernel.cache.SingleVMPool;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.resiliency.spi.cache.SPIPortalCacheManagerConfigurator;
 
 import java.io.Serializable;
 
@@ -32,8 +35,8 @@ import org.osgi.service.component.annotations.Reference;
  * @author Brian Wing Shun Chan
  * @author Michael Young
  */
-@Component(immediate = true, service = SingleVMPool.class)
-public class SingleVMPoolImpl implements SingleVMPool {
+@Component(immediate = true, service = MultiVMPool.class)
+public class MultiVMPoolImpl implements MultiVMPool {
 
 	@Override
 	public void clear() {
@@ -41,21 +44,23 @@ public class SingleVMPoolImpl implements SingleVMPool {
 	}
 
 	@Override
-	public PortalCache<? extends Serializable, ?> getCache(
+	public PortalCache<? extends Serializable, ? extends Serializable> getCache(
 		String portalCacheName) {
 
 		return _portalCacheManager.getCache(portalCacheName);
 	}
 
 	@Override
-	public PortalCache<? extends Serializable, ?> getCache(
+	public PortalCache<? extends Serializable, ? extends Serializable> getCache(
 		String portalCacheName, boolean blocking) {
 
 		return _portalCacheManager.getCache(portalCacheName, blocking);
 	}
 
 	@Override
-	public PortalCacheManager<? extends Serializable, ?> getCacheManager() {
+	public PortalCacheManager<? extends Serializable, ? extends Serializable>
+		getCacheManager() {
+
 		return _portalCacheManager;
 	}
 
@@ -67,6 +72,17 @@ public class SingleVMPoolImpl implements SingleVMPool {
 	@Activate
 	@Modified
 	protected void activate(Map<String, Object> properties) {
+		try {
+			_portalCacheManager =
+				_spiPortalCacheManagerConfigurator.createSPIPortalCacheManager(
+					_portalCacheManager);
+		}
+		catch (Exception e) {
+			if (_log.isErrorEnabled()) {
+				_log.error("Unable to create SPI portal cache manager ", e);
+			}
+		}
+
 		_portalCacheManager.clearAll();
 	}
 
@@ -77,8 +93,8 @@ public class SingleVMPoolImpl implements SingleVMPool {
 
 	@Reference(
 		target =
-			"portal.cache.manager.bridge=" +
-				"com.liferay.portal.cache.SingleVMPortalCacheManagerBridge",
+			"(portal.cache.manager.bridge=" +
+				"com.liferay.portal.cache.internal.MultiVMPortalCacheManagerBridge)",
 		unbind = "-"
 	)
 	protected void setPortalCacheManager(
@@ -88,6 +104,19 @@ public class SingleVMPoolImpl implements SingleVMPool {
 		_portalCacheManager = portalCacheManager;
 	}
 
-	private PortalCacheManager<? extends Serializable, ?> _portalCacheManager;
+	@Reference(unbind = "-")
+	protected void setSPIPortalCacheManagerConfigurator(
+		SPIPortalCacheManagerConfigurator spiPortalCacheManagerConfigurator) {
+
+		_spiPortalCacheManagerConfigurator = spiPortalCacheManagerConfigurator;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		MultiVMPoolImpl.class);
+
+	private PortalCacheManager<? extends Serializable, ? extends Serializable>
+		_portalCacheManager;
+	private SPIPortalCacheManagerConfigurator
+		_spiPortalCacheManagerConfigurator;
 
 }
